Fix branded checkout thank you styling by refactoring thank you component
- Create thankYouSummary component to hold the main content of the thank you page to be reusable in branded checkout
- Update Readme with branded checkout instructions and fixed outdated usage docs
- Move account benefits modal code from thankYou component to accountBenefits component
- Make error loading saved payments alert a warning instead of error
