+++ src/app/checkout/step-2/step-2.tpl.html
@@ -1,4 +1,4 @@
-<div class="alert alert-danger" role="alert" ng-if="$ctrl.loadingExistingPaymentError">
+<div class="alert alert-warning" role="alert" ng-if="$ctrl.loadingExistingPaymentError">
   <p ng-if="$ctrl.loadingExistingPaymentError" ng-switch="$ctrl.loadingExistingPaymentError">
     <span ng-switch-default translate>
       We ran into a problem trying to load your saved payment methods. Please enter your payment details below. Sorry for the inconvenience.
