+++ src/app/thankYou/summary/thankYouSummary.tpl.html
@@ -1,0 +1,155 @@
+<div class="panel loading-overlay-parent">
+  <div class="panel-body">
+    <h2 class="panel-name" translate>
+      Thank you for Your Gift
+    </h2>
+
+    <loading type="overlay" ng-if="$ctrl.loading"></loading>
+    <div class="alert alert-danger" role="alert" ng-if="$ctrl.loadingError" ng-switch="$ctrl.loadingError">
+      <p ng-switch-when="lastPurchaseLink missing" translate>
+        Your gift was submitted but we cannot provide a detailed summary at this time as your session has expired.
+        You will receive a confirmation email from us shortly, or you can review your donation now by looking at <a href="/your-giving.html">Your Giving</a>.
+      </p>
+      <p ng-switch-when="api error" translate>
+        Your gift was submitted but we cannot provide a detailed summary at this time.
+        You will receive a confirmation email from us shortly, or you can review your donation now by looking at <a href="/your-giving.html">Your Giving</a>.
+      </p>
+    </div>
+
+    <p translate>
+      We are processing your gift now. We'll send an email confirmation to
+      <strong>{{$ctrl.email}}</strong> as well as a tax-deductible receipt by postal mail.
+    </p>
+
+    <p translate>
+      Thanks for making a difference in the lives of people both now and for eternity. May the Lord bless your investment in this work for Christ's kingdom. (Luke 6:38)
+    </p>
+
+    <div ng-if="!$ctrl.loading && !$ctrl.loadingError">
+      <div class="row mt">
+        <div class="col-md-6">
+          <div class="panel panel-default">
+            <div class="panel-heading" translate>
+              Payment Method
+            </div>
+
+            <div class="panel-body" ng-if="$ctrl.purchase.paymentMeans.self.type === 'elasticpath.bankaccountpurchases.payment-means-bank-account'">
+              <div class="form-group">
+                <label translate>Bank Name</label>
+                <p>{{$ctrl.purchase.paymentMeans['bank-name']}}</p>
+              </div>
+
+              <div class="form-group">
+                <label translate>Account Type</label>
+                <p>{{$ctrl.purchase.paymentMeans['account-type'] | capitalize}}</p>
+              </div>
+
+              <div class="form-group">
+                <label translate>Routing Number</label>
+                <p>{{$ctrl.purchase.paymentMeans['routing-number']}}</p>
+              </div>
+
+              <div class="form-group">
+                <label translate>Account Number</label>
+                <p>************{{$ctrl.purchase.paymentMeans['display-account-number']}}</p>
+              </div>
+            </div>
+
+            <div class="panel-body" ng-if="$ctrl.purchase.paymentMeans.self.type === 'elasticpath.purchases.purchase.paymentmeans'">
+              <div class="form-group">
+                <label translate>Name on card</label>
+                <p>{{$ctrl.purchase.paymentMeans['holder-name']}}</p>
+              </div>
+
+              <div class="form-group">
+                <label translate>Card Number</label>
+                <p>************{{$ctrl.purchase.paymentMeans['primary-account-number-id']}}</p>
+              </div>
+
+              <div class="form-group">
+                <label translate>Card Type</label>
+                <p>{{$ctrl.purchase.paymentMeans['card-type']}}</p>
+              </div>
+
+              <div class="form-group">
+                <label translate>Expires</label>
+                <p>{{$ctrl.purchase.paymentMeans['expiry-date'].month}}/{{$ctrl.purchase.paymentMeans['expiry-date'].year}}</p><!-- TODO: deal with localization -->
+              </div>
+
+              <div class="form-group">
+                <label translate>Billing Address</label>
+                <display-address address="$ctrl.purchase.paymentMeans.address"></display-address>
+              </div>
+            </div>
+          </div>
+        </div>
+        <div class="col-md-6">
+          <div class="panel panel-default">
+            <div class="panel-heading" translate>
+              Receipt Mailing Address
+            </div>
+
+            <div class="panel-body">
+              <div class="form-group">
+                <div>
+                  {{$ctrl.purchase.donorDetails.name.title}}
+                  {{$ctrl.purchase.donorDetails.name['given-name']}}
+                  {{$ctrl.purchase.donorDetails.name['middle-initial']}}
+                  {{$ctrl.purchase.donorDetails.name['family-name']}}
+                  {{$ctrl.purchase.donorDetails.name.suffix}}
+                </div>
+                <div ng-if="$ctrl.purchase.donorDetails['donor-type'] === 'Household'">
+                  {{$ctrl.purchase.donorDetails['spouse-name'].title}}
+                  {{$ctrl.purchase.donorDetails['spouse-name']['given-name']}}
+                  {{$ctrl.purchase.donorDetails['spouse-name']['middle-initial']}}
+                  {{$ctrl.purchase.donorDetails['spouse-name']['family-name']}}
+                  {{$ctrl.purchase.donorDetails['spouse-name'].suffix}}
+                </div>
+                <div ng-if="$ctrl.purchase.donorDetails['donor-type'] === 'Organization'">
+                  {{$ctrl.purchase.donorDetails['organization-name']}}
+                </div>
+                <display-address address="$ctrl.purchase.donorDetails.mailingAddress"></display-address>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div><!-- // row -->
+
+      <div class="row">
+        <div class="col-md-12">
+          <div class="panel panel-default">
+            <div class="panel-body">
+              <h4 class="panel-name" translate>Gift Summary</h4>
+
+              <table class="table giftsum-table giftsum-table-cart">
+                <tbody>
+                <tr class="giftsum-gift-row" ng-repeat="item in $ctrl.purchase.lineItems">
+                  <td class="td-gift">
+                    <span class="giftsum-person giftsum-title">{{item.name}}</span>
+                    <p>#{{item.code['product-code']}}</p>
+                  </td>
+                  <td class="td-frequency">
+                    <span class="giftsum-about giftsum-title">{{item.rate.display}}</span>
+                    <span class="giftsum-frequency giftsum-title">
+                                <span ng-if="item.rate.recurrence.interval !== 'NA'" translate>First Gift: {{item.rate['start-date'].value | date:'mediumDate':'UTC'}}</span>
+                              </span>
+                  </td>
+                </tr>
+                </tbody>
+
+                <tfoot>
+                <tr>
+                  <td class="text-right" colspan="2">
+                    <display-rate-totals rate-totals="$ctrl.rateTotals"></display-rate-totals>
+                  </td>
+                </tr>
+                </tfoot>
+              </table>
+            </div>
+          </div>
+        </div>
+      </div><!-- // row -->
+    </div>
+
+  </div>
+</div><!-- // panel -->
